refactor: migrate logo components
