+++ src/app/(auth)/login/page.tsx
@@ -4,7 +4,7 @@ import { redirect } from 'next/navigation';
 
 import EmailLogIn from '@/components/auth/EmailLogInForm';
 import OauthLogIn from '@/components/auth/OauthLogInForm';
-import LogoColored from '@/components/ui/layout/LogoColored';
+import LogoColored from '@/components/icons/LogoColored';
 
 import { getUserBySession } from '@/lib/api/user';
