+++ src/components/icons/LogoColored.tsx
