+++ src/components/icons/LogoWhite.tsx
